requirements.txt
tensorflow
numpy
opencv-python
matplotlib
streamlit
pillow
scikit-learn
